pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.1",
    "tgcrypto>=1.2.5",
    "psutil>=5.9.0",
]
